+++ src/Command/MakeDefaultsCommand.php
@@ -22,8 +22,9 @@ class MakeDefaultsCommand extends AbstractMakeCommand
     {
         $this
             ->setDescription('Creates a new PHP class for the default Autotask services.')
-            ->addOption('output', 'o', InputOption::VALUE_REQUIRED, 'Sets the output directory for the generated classes.')
-            ->addOption('force', 'f', InputOption::VALUE_NONE, 'Overwrites previous classes if they exist.');
+            ->addOption('no-cache', null, InputOption::VALUE_NONE, 'Specifies whether any previous cached files should be ignored (still makes use of cache for efficiency, but clears afterward).')
+            ->addOption('force', 'f', InputOption::VALUE_NONE, 'Overwrites previous classes if they exist.')
+            ->addOption('output', 'o', InputOption::VALUE_REQUIRED, 'Sets the output directory for the generated classes.');
 
     }
 
@@ -33,20 +34,12 @@ class MakeDefaultsCommand extends AbstractMakeCommand
     protected function execute(InputInterface $input, OutputInterface $output)
     {
         $this->setupEnvironment($input, $output);
-        $endpoint = $this->generator->endpoint();
+        $endpoints = Api::endpoints();
 
         try {
-            $outputDirectory = $input->getOption('output') ?? Env::get('AUTOTASK_GENERATOR_DIRECTORY', getcwd());
-            $endpoint->setOutputDirectory($outputDirectory);
-
-            if ($input->getOption('force')) {
-                $endpoint->setOverwrite(true);
-            }
-
-            foreach (Api::endpoints() as $service) {
+            foreach ($endpoints as $service) {
                 $output->writeln('Generating classes for ' . $service);
-                $endpoint->setEndpoint($service);
-                $endpoint->make();
+                $this->generator->makeResource($service);
             }
         } catch (\Exception $error) {
             $output->writeln(
@@ -57,13 +50,20 @@ class MakeDefaultsCommand extends AbstractMakeCommand
             return Command::FAILURE;
         }
 
-        $output->writeln('<info>Successfully created endpoints!');
+        $output->writeln(
+            '<info>Successfully created '.
+            count($endpoints).
+            ' endpoints in '.
+            (microtime(true) - $_SERVER['REQUEST_TIME_FLOAT']).
+            ' seconds!</info>'
+        );
+        $output->writeln('');
 
         // Regenerate support classes
         $command = $this->getApplication()->find('make:support-files');
         return $command->run(
             new ArrayInput([
-                '-o' => $outputDirectory,
+                '-o' => $input->getOption('output') ?? Env::get('AUTOTASK_GENERATOR_DIRECTORY', getcwd()),
                 '-f' => true,
             ]),
             $output